Basic moving around of assets.
