+++ packages/backend-core/package.json
@@ -60,6 +60,7 @@
     "@swc/core": "^1.3.25",
     "@swc/jest": "^0.2.24",
     "@trendyol/jest-testcontainers": "^2.1.1",
+    "@jest/test-sequencer": "29.5.0",
     "@types/chance": "1.1.3",
     "@types/ioredis": "4.28.0",
     "@types/jest": "28.1.1",
@@ -76,13 +77,13 @@
     "@types/uuid": "8.3.4",
     "chance": "1.1.8",
     "ioredis-mock": "5.8.0",
-    "jest": "28.1.1",
+    "jest": "29.5.0",
     "jest-serial-runner": "^1.2.1",
     "koa": "2.13.4",
     "nodemon": "2.0.16",
     "pouchdb-adapter-memory": "7.2.2",
     "timekeeper": "2.2.0",
-    "ts-jest": "28.0.4",
+    "ts-jest": "29.0.5",
     "ts-node": "10.8.1",
     "tsconfig-paths": "4.0.0",
     "typescript": "4.7.3"
