+++ packages/backend-core/src/db/documents/index.ts
@@ -1,0 +1,2 @@
+export * from "./ids"
+export * from "./params"
