+++ packages/backend-core/tests/core/utilities/index.ts
@@ -1,9 +1,6 @@
 export * as mocks from "./mocks"
 export * as structures from "./structures"
 export { generator } from "./structures"
-export * as testEnv from "./testEnv"
 export * as testContainerUtils from "./testContainerUtils"
 
 export * from "./jestUtils"
-
-export { default as DBTestConfiguration } from "./DBTestConfiguration"
