+++ packages/backend-core/tests/core/utilities/mocks/events.ts
@@ -1,0 +1,1 @@
+jest.mock("../../../src/events")
