+++ packages/backend-core/tests/extra/index.ts
@@ -1,0 +1,2 @@
+export * as testEnv from "./testEnv"
+export * as DBTestConfiguration from "./DBTestConfiguration"
